Updating TabBar tokenization  with new compound control strategy
